feat: moved useData hook to own file
